refactor to a hook. i want to use this in other things. 🤓
